+++ linkerd/core/src/test/scala/io/buoyant/linkerd/HostConnectionPoolTest.scala
@@ -2,7 +2,9 @@ package io.buoyant.linkerd
 
 import com.twitter.conversions.time._
 import com.twitter.finagle.client.DefaultPool
+import com.twitter.finagle.Path
 import io.buoyant.linkerd.Linker.Initializers
+import io.buoyant.router.StackRouter.Client.PerClientParams
 import org.scalatest.FunSuite
 
 class HostConnectionPoolTest extends FunSuite {
@@ -22,7 +24,8 @@ class HostConnectionPoolTest extends FunSuite {
                    |""".stripMargin
 
     val linker = Linker.load(config, Initializers(protocol = Seq(TestProtocol.Plain)))
-    val pool = linker.routers.head.params[DefaultPool.Param]
+    val params = linker.routers.head.params[PerClientParams].paramsFor(Path.read("/foo"))
+    val pool = params[DefaultPool.Param]
     assert(pool.low == 5)
     assert(pool.high == 100)
     assert(pool.idleTime == 5.seconds)